moving skeletons to separate folder
