+++ src/components/Hero/Benchmark.jsx
@@ -1,30 +1,14 @@
 import React, { Component } from 'react';
 import { shape, func, bool, arrayOf, oneOfType, string } from 'prop-types';
 import { connect } from 'react-redux';
-import ContentLoader from 'react-content-loader';
 import { getBenchmark } from '../../actions';
 import BenchmarkTable from './BenchmarkTable';
+import BenchmarkSkeleton from '../Skeletons/BenchmarkSkeleton';
 
 const renderBenchmark = (hero, data) => (
   <div>
     <BenchmarkTable data={data} />
   </div>
-);
-
-const BenchmarkSkeleton = props => (
-  <ContentLoader
-    primaryColor="#371b68"
-    secondaryColor="#371b68"
-    width={400}
-    animate={false}
-    {...props}
-  >
-    <rect x="0" y="10" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="25" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="40" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="55" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="70" rx="5" ry="5" width="300" height="5" />
-  </ContentLoader>
 );
 
 class Benchmark extends Component {
