+++ src/components/Hero/Players.jsx
@@ -1,27 +1,11 @@
 import React from 'react';
 import { bool, func, arrayOf, shape, number, string } from 'prop-types';
 import { connect } from 'react-redux';
-import ContentLoader from 'react-content-loader';
 import { getHeroPlayers } from '../../actions';
 import Table, { TableLink } from '../Table';
+import PlayersSkeleton from '../Skeletons/PlayersSkeleton';
 import { wilsonScore } from '../../utility';
 import { proPlayersSelector } from '../../reducers/selectors';
-
-const PlayersSkeleton = props => (
-  <ContentLoader
-    primaryColor="#371b68"
-    secondaryColor="#371b68"
-    width={400}
-    animate={false}
-    {...props}
-  >
-    <rect x="0" y="10" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="25" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="40" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="55" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="70" rx="5" ry="5" width="300" height="5" />
-  </ContentLoader>
-);
 
 class Players extends React.Component {
   static propTypes = {
