+++ src/components/Hero/Recent.jsx
@@ -1,28 +1,12 @@
 import React from 'react';
 import { arrayOf, shape, bool, func, string, oneOfType } from 'prop-types';
 import { connect } from 'react-redux';
-import ContentLoader from 'react-content-loader';
 import Table, { TableLink } from '../Table';
+import RecentSkeleton from '../Skeletons/RecentSkeleton'
 import ErrorBox from '../Error/ErrorBox';
 import { getHeroRecentGames } from '../../actions';
 import { transformations } from '../../utility';
 import { proPlayersSelector } from '../../reducers/selectors';
-
-const RecentSkeleton = props => (
-  <ContentLoader
-    primaryColor="#371b68"
-    secondaryColor="#371b68"
-    width={400}
-    animate={false}
-    {...props}
-  >
-    <rect x="0" y="10" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="25" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="40" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="55" rx="5" ry="5" width="300" height="5" />
-    <rect x="0" y="70" rx="5" ry="5" width="300" height="5" />
-  </ContentLoader>
-);
 
 class Recent extends React.Component {
   static propTypes = {
